Move additional helpers that use `pandas` to `pandas_adjuster.py`.
These will be removed later, but moving them to a single file will make that
cleaner.
